refactor(#52): move everything to two packages core , api
