+++ src/main/java/com/jrest/java/api/annotation/REQUEST.java
@@ -1,4 +1,4 @@
-package com.jrest.java.annotation;
+package com.jrest.java.api.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,7 +6,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import com.jrest.java.util.HTTP_METHOD;
+import com.jrest.java.api.HTTP_METHOD;
 
 /**
  * Use this mandatory annotation to specify the HTTP Method type
