+++ src/main/java/com/jrest/java/core/APIClient.java
@@ -1,6 +1,7 @@
-package com.jrest.java.util;
+package com.jrest.java.core;
 
-import com.jrest.java.model.APIRequestRecord;
+import com.jrest.java.api.APIResponse;
+
 //TODO: Add documentation
 public interface APIClient {
     <Response> APIResponse<Response> fetch(final APIRequestRecord apiRequestRecord) throws Exception;
