+++ src/main/java/com/jrest/java/core/RequestProxyRecord.java
@@ -1,4 +1,4 @@
-package com.jrest.java.model;
+package com.jrest.java.core;
 
 public record RequestProxyRecord(
         String url,
