+++ src/test/java/com/jrest/java/AsyncTest.java
@@ -1,8 +1,8 @@
 package com.jrest.java;
 
-import com.jrest.java.util.APICallBack;
-import com.jrest.java.util.APIResponse;
-import com.jrest.java.util.JRest;
+import com.jrest.java.api.APICallBack;
+import com.jrest.java.api.APIResponse;
+import com.jrest.java.api.JRest;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
